[ENH] refactoring entity structure into ORM module
