--- src/Core/BaseEntityModel/Provider/Contract/EntityInterface.php
@@ -1,6 +1,0 @@
-<?php
-namespace Wepesi\Core\BaseEntityModel\Provider\Contract;
-
-interface EntityInterface {
-    public function findAll():array;
-}
