+++ src/Core/Orm/EntityModel/Provider/Contract/EntityInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace Wepesi\Core\Orm\EntityModel\Provider\Contract;
+
+interface EntityInterface {
+    public function findAll():array;
+    public function findOne():array;
+}
