+++ src/Core/Orm/Relations/HasMany.php
@@ -2,7 +2,7 @@
 
 namespace Wepesi\Core\Orm\Relations;
 
-use Wepesi\Core\BaseEntityModel\Provider\Contract\EntityInterface;
+use Wepesi\Core\Orm\EntityModel\Provider\Contract\EntityInterface;
 
 
 /**
